sender/sender/users/api/serializers.py
from django.contrib.auth import get_user_model
from rest_framework import serializers

from sender.users.models import Contacts

User = get_user_model()


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ["username", "name", "url"]

        extra_kwargs = {
            "url": {"view_name": "api:user-detail", "lookup_field": "username"},
        }

class ContactsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contacts
        fields = ['from_who', 'to', 'text']
